Remove old go back button
