+++ front/app/containers/Admin/pagesAndMenu/containers/ContentBuilder/components/HomepageBuilderTopBar/index.tsx
@@ -7,7 +7,7 @@ import { SupportedLocale } from 'typings';
 import useAddHomepageBuilderLayout from 'api/home_page_layout/useAddHomepageLayout';
 
 import Container from 'components/admin/ContentBuilder/TopBar/Container';
-import GoBackButton from 'components/admin/ContentBuilder/TopBar/GoBackButton2';
+import GoBackButton from 'components/admin/ContentBuilder/TopBar/GoBackButton';
 import LocaleSwitcher from 'components/admin/ContentBuilder/TopBar/LocaleSwitcher';
 import PreviewToggle from 'components/admin/ContentBuilder/TopBar/PreviewToggle';
 import SaveButton from 'components/admin/ContentBuilder/TopBar/SaveButton';
